chore: move save_dp_model and load_dp_model to a seperated module
Fix #3526.
